// frontend/src/app/api/computer-use/actions.ts
export enum BrowserActions {
    MOUSE_CLICK = "mouseClick",
    CLICK_ELEMENT = "click",
    FILL_INPUT = "fill",
    PRESS = "press",
    NAVIGATE_TO = "navigateTo",
    SCROLL_TO = "scrollTo",
    EXTRACT = "extract",
    BACK = "back",
    FORWARD = "forward",
    REFRESH = "refresh",
    RELOAD = "reload",
    FILL_LOGIN = "fillLogin",
}